Add unit tests for ChannelFrame processing and transformation methods
- Implemented comprehensive tests for various methods in ChannelFrame, including high-pass and low-pass filters, A-weighting, absolute value, power, summation, and mean calculations.
- Added tests for channel difference, trimming, HPSS operations, and addition with SNR.
- Created tests for transformations such as FFT, Welch, STFT, N-octave spectrum, cross-spectral density, transfer function, and coherence.
- Ensured that all operations are lazy and do not trigger computation until explicitly called.
- Verified the correctness of outputs and maintained proper channel metadata throughout the tests.
