+++ tests/frames/test_channel_frame.py
@@ -131,51 +131,6 @@ class TestChannelFrame:
             _: NDArrayReal = channel.data
             mock_compute.assert_called_once()
 
-    def test_filter_operations(self) -> None:
-        """Test that filter operations are lazy."""
-        with mock.patch("wandas.processing.create_operation") as mock_create_op:
-            mock_op: mock.MagicMock = mock.MagicMock()
-            mock_op.process.return_value = self.dask_data
-            mock_create_op.return_value = mock_op
-
-            # Apply filter operations
-            result: ChannelFrame = self.channel_frame.high_pass_filter(cutoff=100)
-            mock_create_op.assert_called_with(
-                "highpass_filter", self.sample_rate, cutoff=100, order=4
-            )
-
-            result = self.channel_frame.low_pass_filter(cutoff=5000)
-            mock_create_op.assert_called_with(
-                "lowpass_filter", self.sample_rate, cutoff=5000, order=4
-            )
-
-            # No compute should have happened
-            mock_op.process.assert_called()
-            assert isinstance(result, ChannelFrame)
-
-        with mock.patch.object(
-            DaArray, "compute", return_value=self.dask_data.compute()
-        ) as mock_compute:
-            # Apply filter operations
-            result = self.channel_frame.high_pass_filter(cutoff=100)
-            mock_compute.assert_not_called()
-
-            result = self.channel_frame.low_pass_filter(cutoff=5000)
-            mock_compute.assert_not_called()
-
-            # Check that the result has the expected type
-            assert isinstance(result, ChannelFrame)
-            assert isinstance(result._data, DaArray)
-            assert result.n_channels == 2
-            assert result.n_samples == 16000
-            assert result.sampling_rate == self.sample_rate
-            assert result.label == "test_audio"
-            assert result.channels[0].label == "ch0"
-            assert result.channels[1].label == "ch1"
-            assert result.shape == (2, 16000)
-            assert result.data.shape == (2, 16000)
-            np.testing.assert_array_equal(result.data, self.data)
-
     def test_plotting_triggers_compute(self) -> None:
         """Test that plotting triggers computation."""
         with mock.patch("wandas.frames.channel.create_operation") as mock_get_strategy:
@@ -401,250 +356,50 @@ class TestChannelFrame:
         ):
             _ = self.channel_frame + other_cf
 
-    def test_sum_methods(self) -> None:
-        """Test sum() methods."""
-        # Test that sum method is lazy
-        with mock.patch.object(DaArray, "compute") as mock_compute:
-            # Call sum() - this should be lazy and not trigger computation
-            sum_cf = self.channel_frame.sum()
+    def test_add_method(self) -> None:
+        """Test add method for adding signals."""
+        # 通常の加算をテスト
+        # Create another ChannelFrame
+        other_data = np.random.random((2, 16000))
+        other_dask_data = _da_from_array(other_data, chunks=-1)
+        other_cf = ChannelFrame(other_dask_data, self.sample_rate, label="other_audio")
 
-            # Check no computation happened yet
-            mock_compute.assert_not_called()
+        # addメソッドを使用して加算
+        result = self.channel_frame.add(other_cf)
 
-            # Verify result is the expected type
-            assert isinstance(sum_cf, ChannelFrame)
-            assert sum_cf.n_channels == 1
+        # Check result properties
+        assert isinstance(result, ChannelFrame)
+        assert result.sampling_rate == self.sample_rate
+        assert result.n_channels == 2
+        assert result.n_samples == 16000
 
-        # Test correctness of computation result
-        sum_cf = self.channel_frame.sum()
-        sum_data = sum_cf.compute()
-        expected_sum = self.data.sum(axis=-2, keepdims=True)
-        np.testing.assert_array_almost_equal(sum_data, expected_sum)
-
-    def test_mean_methods(self) -> None:
-        """Test mean() methods."""
-
-        # Test mean method
-        with mock.patch.object(DaArray, "compute") as mock_compute:
-            # Call sum() - this should be lazy and not trigger computation
-            mean_cf = self.channel_frame.mean()
-
-            # Check no computation happened yet
-            mock_compute.assert_not_called()
-
-            # Verify result is the expected type
-            assert isinstance(mean_cf, ChannelFrame)
-            assert mean_cf.n_channels == 1
-
-        # Compute and check results
-        mean_data = mean_cf.compute()
-        expected_mean = self.data.mean(axis=-2, keepdims=True)
-        np.testing.assert_array_almost_equal(mean_data, expected_mean)
-
-    def test_channel_difference(self) -> None:
-        """Test channel_difference method."""
-        # Test that channel_difference is lazy
-        with mock.patch.object(DaArray, "compute") as mock_compute:
-            # Call channel_difference - this should be lazy and not trigger computation
-            diff_cf = self.channel_frame.channel_difference(other_channel=0)
-
-            # Check no computation happened yet
-            mock_compute.assert_not_called()
-
-            # Verify result is the expected type
-            assert isinstance(diff_cf, ChannelFrame)
-            assert diff_cf.n_channels == self.channel_frame.n_channels
-
-        # Test correctness of computation result
-        diff_cf = self.channel_frame.channel_difference(other_channel=0)
-        computed = diff_cf.compute()
-        expected = self.data - self.data[0:1]
+        # 実際の計算結果を確認
+        computed = result.compute()
+        expected = self.data + other_data
         np.testing.assert_array_almost_equal(computed, expected)
 
-        # Test that channel_difference with other_channel=0 works correctly
-        diff_cf = self.channel_frame.channel_difference(other_channel="ch0")
-        computed = diff_cf.compute()
-        expected = self.data - self.data[0:1]
+        # スカラー値との加算をテスト
+        scalar_value = 0.5
+        result = self.channel_frame.add(scalar_value)
+        assert isinstance(result, ChannelFrame)
+        computed = result.compute()
+        expected = self.data + scalar_value
         np.testing.assert_array_almost_equal(computed, expected)
 
-        # Test invalid channel index
-        with pytest.raises(IndexError):
-            self.channel_frame.channel_difference(other_channel=10)
+        # NumPy配列との加算をテスト
+        array_value = np.random.random((2, 16000))
+        result = self.channel_frame.add(array_value)
+        assert isinstance(result, ChannelFrame)
+        computed = result.compute()
+        expected = self.data + array_value
+        np.testing.assert_array_almost_equal(computed, expected)
 
-    def test_additional_filter_operations(self) -> None:
-        """Test normalize and a_weighting operations."""
-        with mock.patch("wandas.processing.create_operation") as mock_create_op:
-            mock_op = mock.MagicMock()
-            mock_op.process.return_value = self.dask_data
-            mock_create_op.return_value = mock_op
-
-            # Test normalize
-            # result = self.channel_frame.normalize(
-            # target_level=-15, channel_wise=False)
-            # mock_create_op.assert_called_with(
-            #     "normalize", self.sample_rate, target_level=-15, channel_wise=False
-            # )
-            # assert isinstance(result, ChannelFrame)
-
-            # Test a_weighting
-            result = self.channel_frame.a_weighting()
-            mock_create_op.assert_called_with("a_weighting", self.sample_rate)
-            assert isinstance(result, ChannelFrame)
-
-            # Test HPSS methods
-            result = self.channel_frame.hpss_harmonic(kernel_size=31)
-            mock_create_op.assert_called_with(
-                "hpss_harmonic",
-                self.sample_rate,
-                kernel_size=31,
-                power=2,
-                margin=1,
-                n_fft=2048,
-                hop_length=None,
-                win_length=None,
-                window="hann",
-                center=True,
-                pad_mode="constant",
-            )
-            assert isinstance(result, ChannelFrame)
-
-            result = self.channel_frame.hpss_percussive(kernel_size=31)
-            mock_create_op.assert_called_with(
-                "hpss_percussive",
-                self.sample_rate,
-                kernel_size=31,
-                power=2,
-                margin=1,
-                n_fft=2048,
-                hop_length=None,
-                win_length=None,
-                window="hann",
-                center=True,
-                pad_mode="constant",
-            )
-            assert isinstance(result, ChannelFrame)
-
-    def test_visualize_graph(self) -> None:
-        """Test visualize_graph method."""
-        # Test successful visualization
-        with mock.patch.object(DaArray, "visualize") as mock_visualize:
-            filename = self.channel_frame.visualize_graph()
-            mock_visualize.assert_called_once()
-            assert filename is not None
-
-        # Test with provided filename
-        with mock.patch.object(DaArray, "visualize") as mock_visualize:
-            custom_filename = "test_graph.png"
-            filename = self.channel_frame.visualize_graph(filename=custom_filename)
-            mock_visualize.assert_called_with(filename=custom_filename)
-            assert filename == custom_filename
-
-        # Test handling of visualization error
-        with mock.patch.object(
-            DaArray, "visualize", side_effect=Exception("Test error")
+        # サンプリングレートが一致しない場合のエラーをテスト
+        mismatch_cf = ChannelFrame(other_dask_data, 44100, label="mismatch_audio")
+        with pytest.raises(
+            ValueError, match="Sampling rates do not match. Cannot perform operation."
         ):
-            filename = self.channel_frame.visualize_graph()
-            assert filename is None
-
-    def test_read_wav_class_method(self) -> None:
-        """Test read_wav class method."""
-        with mock.patch.object(ChannelFrame, "from_file") as mock_from_file:
-            mock_from_file.return_value = self.channel_frame
-            result = ChannelFrame.read_wav("test.wav", labels=["left", "right"])
-            mock_from_file.assert_called_with("test.wav", ch_labels=["left", "right"])
-            assert result is self.channel_frame
-
-    def test_debug_info(self) -> None:
-        """Test debug_info method."""
-        with mock.patch("wandas.core.base_frame.logger") as mock_logger:
-            self.channel_frame.debug_info()
-            assert mock_logger.debug.call_count >= 6  # At least 9 debug messages
-
-    @pytest.mark.integration  # type: ignore [misc, unused-ignore]
-    def test_from_file_lazy_loading(self) -> None:
-        """Test that loading from file is lazy."""
-        # Create a temporary WAV file
-        with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as temp_file:
-            temp_filename: str = temp_file.name
-            sf.write(temp_filename, np.random.random((16000, 2)), 16000)
-        re_test_data, _ = sf.read(temp_filename)
-        re_test_data = re_test_data.T
-        try:
-            # Create mock array and patch from_delayed to return it
-            mock_dask_array = mock.MagicMock(spec=DaArray)
-            mock_data = np.random.random((2, 16000))
-            mock_dask_array.compute.return_value = mock_data
-            mock_dask_array.shape = (2, 16000)
-            # Add ndim property to the mock
-            mock_dask_array.ndim = 2
-
-            # Mock the rechunk method to return the same mock
-            mock_dask_array.rechunk.return_value = mock_dask_array
-
-            # Patch necessary functions
-            with (
-                mock.patch("wandas.io.readers.get_file_reader") as mock_get_reader,
-                mock.patch("dask.array.from_delayed", return_value=mock_dask_array),
-                mock.patch("dask.delayed", return_value=mock.MagicMock()),
-            ):
-                # Set up mock reader
-                mock_reader = mock.MagicMock()
-                mock_reader.get_file_info.return_value = {
-                    "samplerate": 16000,
-                    "channels": 2,
-                    "frames": 16000,
-                }
-                mock_reader.get_audio_data.return_value = mock_data
-                mock_get_reader.return_value = mock_reader
-
-                # Create ChannelFrame from file
-                cf: ChannelFrame = ChannelFrame.from_file(temp_filename)
-
-                # Check file reading hasn't happened yet
-                mock_reader.get_audio_data.assert_not_called()
-
-                # Access data to trigger computation
-                data: NDArrayReal = cf.data
-
-                # Verify data is correct
-                np.testing.assert_array_equal(data, re_test_data)
-
-                # Test with channel selection parameters
-                cf = ChannelFrame.from_file(
-                    temp_filename, channel=0, start=0.1, end=0.5
-                )
-                # assert cf.metadata["channels"] == [0]
-                assert cf.channels[0].label == "ch0"
-                # Test with multiple channels
-                cf = ChannelFrame.from_file(temp_filename, channel=[0, 1])
-                # assert cf.metadata["channels"] == [0, 1]
-                assert cf.channels[0].label == "ch0"
-                assert cf.channels[1].label == "ch1"
-                # Test error cases
-                with pytest.raises(
-                    ValueError, match="Channel specification is out of range"
-                ):
-                    ChannelFrame.from_file(temp_filename, channel=5)
-
-                with pytest.raises(
-                    ValueError, match="Channel specification is out of range"
-                ):
-                    ChannelFrame.from_file(temp_filename, channel=[0, 5])
-
-                with pytest.raises(
-                    TypeError,
-                    match="channel must be int, list, or None",
-                ):
-                    ChannelFrame.from_file(temp_filename, channel="invalid")  # type: ignore
-
-                # Test file not found
-                with pytest.raises(FileNotFoundError):
-                    ChannelFrame.from_file("nonexistent_file.wav")
-
-        finally:
-            # Clean up the temporary file
-            if os.path.exists(temp_filename):
-                os.remove(temp_filename)
+            _ = self.channel_frame.add(mismatch_cf)
 
     def test_channel_metadata_label_access(self) -> None:
         """Test accessing and modifying channel labels through metadata."""
@@ -783,235 +538,90 @@ class TestChannelFrame:
                 sampling_rate=sampling_rate,
             )
 
-    def test_read_csv(self) -> None:
-        """Test read_csv method."""
-        # Create a temporary CSV file
-        with tempfile.NamedTemporaryFile(suffix=".csv", delete=False) as temp_file:
-            temp_filename = temp_file.name
-            # 時間と2つの値を持つCSVファイルを作成
-            header = "time,value1,value2\n"
-            data = "\n".join([f"{i / 16000},{1.1},{2.2}" for i in range(16000)])
-            temp_file.write(header.encode())
-            temp_file.write(data.encode())
-            temp_file.flush()
-            temp_file.seek(0)
-            # Close the file to ensure it's written
-            temp_file.close()
-
+    def test_from_file_lazy_loading(self) -> None:
+        """Test that loading from file is lazy."""
+        # Create a temporary WAV file
+        with tempfile.NamedTemporaryFile(suffix=".wav", delete=False) as temp_file:
+            temp_filename: str = temp_file.name
+            sf.write(temp_filename, np.random.random((16000, 2)), 16000)
+        re_test_data, _ = sf.read(temp_filename)
+        re_test_data = re_test_data.T
         try:
-            # Read the CSV file into a ChannelFrame
-            cf = ChannelFrame.read_csv(temp_filename)
+            # Create mock array and patch from_delayed to return it
+            mock_dask_array = mock.MagicMock(spec=DaArray)
+            mock_data = np.random.random((2, 16000))
+            mock_dask_array.compute.return_value = mock_data
+            mock_dask_array.shape = (2, 16000)
+            # Add ndim property to the mock
+            mock_dask_array.ndim = 2
 
-            # Check properties
-            assert cf.sampling_rate == 16000
-            assert cf.n_channels == 2
-            assert cf.n_samples == 16000
-            assert cf.label == Path(temp_filename).stem
+            # Mock the rechunk method to return the same mock
+            mock_dask_array.rechunk.return_value = mock_dask_array
 
-            # Check data
-            expected_data = np.loadtxt(temp_filename, delimiter=",", skiprows=1).T
-            np.testing.assert_array_equal(cf.data, expected_data[1:])
+            # Patch necessary functions
+            with (
+                mock.patch("wandas.io.readers.get_file_reader") as mock_get_reader,
+                mock.patch("dask.array.from_delayed", return_value=mock_dask_array),
+                mock.patch("dask.delayed", return_value=mock.MagicMock()),
+            ):
+                # Set up mock reader
+                mock_reader = mock.MagicMock()
+                mock_reader.get_file_info.return_value = {
+                    "samplerate": 16000,
+                    "channels": 2,
+                    "frames": 16000,
+                }
+                mock_reader.get_audio_data.return_value = mock_data
+                mock_get_reader.return_value = mock_reader
+
+                # Create ChannelFrame from file
+                cf: ChannelFrame = ChannelFrame.from_file(temp_filename)
+
+                # Check file reading hasn't happened yet
+                mock_reader.get_audio_data.assert_not_called()
+
+                # Access data to trigger computation
+                data: NDArrayReal = cf.data
+
+                # Verify data is correct
+                np.testing.assert_array_equal(data, re_test_data)
+
+                # Test with channel selection parameters
+                cf = ChannelFrame.from_file(
+                    temp_filename, channel=0, start=0.1, end=0.5
+                )
+                # assert cf.metadata["channels"] == [0]
+                assert cf.channels[0].label == "ch0"
+                # Test with multiple channels
+                cf = ChannelFrame.from_file(temp_filename, channel=[0, 1])
+                # assert cf.metadata["channels"] == [0, 1]
+                assert cf.channels[0].label == "ch0"
+                assert cf.channels[1].label == "ch1"
+                # Test error cases
+                with pytest.raises(
+                    ValueError, match="Channel specification is out of range"
+                ):
+                    ChannelFrame.from_file(temp_filename, channel=5)
+
+                with pytest.raises(
+                    ValueError, match="Channel specification is out of range"
+                ):
+                    ChannelFrame.from_file(temp_filename, channel=[0, 5])
+
+                with pytest.raises(
+                    TypeError,
+                    match="channel must be int, list, or None",
+                ):
+                    ChannelFrame.from_file(temp_filename, channel="invalid")  # type: ignore
+
+                # Test file not found
+                with pytest.raises(FileNotFoundError):
+                    ChannelFrame.from_file("nonexistent_file.wav")
 
         finally:
+            # Clean up the temporary file
             if os.path.exists(temp_filename):
                 os.remove(temp_filename)
-
-    def test_fft_transform(self) -> None:
-        """Test fft method for lazy transformation to frequency domain."""
-        from wandas.frames.spectral import SpectralFrame
-        from wandas.processing import FFT
-
-        with mock.patch("wandas.processing.create_operation") as mock_create_op:
-            # モックFFTオペレーションの設定
-            mock_fft = mock.MagicMock(spec=FFT)
-            mock_fft.n_fft = 4096
-            mock_fft.window = "hann"
-            mock_data = mock.MagicMock(spec=DaArray)
-            mock_data.ndim = 2  # Set ndim property to pass dimension check
-            mock_data.shape = (2, 2049)  # Set appropriate shape for a 2D array
-            mock_fft.process.return_value = mock_data
-            mock_create_op.return_value = mock_fft
-
-            # fftを遅延実行
-            result = self.channel_frame.fft(n_fft=4096, window="hamming")
-
-            # オペレーションが正しく作成されたか確認
-            mock_create_op.assert_called_with(
-                "fft", self.sample_rate, n_fft=4096, window="hamming"
-            )
-
-            # processメソッドが呼び出されたか確認
-            mock_fft.process.assert_called_once_with(self.channel_frame._data)
-
-            # 結果が正しい型か確認
-            assert isinstance(result, SpectralFrame)
-            assert result.n_fft == 4096
-            assert result.window == "hann"
-            assert result.previous is self.channel_frame
-
-    def test_welch_transform(self) -> None:
-        """
-        Test welch method for lazy transformation to frequency domain
-        using Welch's method.
-        """
-        from wandas.frames.spectral import SpectralFrame
-        from wandas.processing import Welch
-
-        with mock.patch("wandas.processing.create_operation") as mock_create_op:
-            # モックWelchオペレーションの設定
-            mock_welch = mock.MagicMock(spec=Welch)
-            mock_welch.n_fft = 2048
-            mock_welch.hop_length = 256
-            mock_welch.win_length = 1024
-            mock_welch.window = "blackman"
-            mock_welch.average = "mean"
-            mock_data = mock.MagicMock(spec=DaArray)
-            mock_data.ndim = 2  # Set ndim property to pass dimension check
-            mock_data.shape = (2, 1025)  # Set appropriate shape for a 2D array
-            mock_welch.process.return_value = mock_data
-            mock_create_op.return_value = mock_welch
-
-            # welchを遅延実行
-            result = self.channel_frame.welch(
-                n_fft=2048,
-                hop_length=256,
-                win_length=1024,
-                window="blackman",
-                average="mean",
-            )
-
-            # オペレーションが正しく作成されたか確認
-            mock_create_op.assert_called_with(
-                "welch",
-                self.sample_rate,
-                n_fft=2048,
-                hop_length=256,
-                win_length=1024,
-                window="blackman",
-                average="mean",
-            )
-
-            # processメソッドが呼び出されたか確認
-            mock_welch.process.assert_called_once_with(self.channel_frame._data)
-
-            # 結果が正しい型か確認
-            assert isinstance(result, SpectralFrame)
-            assert result.n_fft == 2048
-            assert result.window == "blackman"
-            assert result.previous is self.channel_frame
-
-    def test_stft_transform(self) -> None:
-        """Test stft method for lazy short-time Fourier transform."""
-        from wandas.frames.spectrogram import SpectrogramFrame
-        from wandas.processing import STFT
-
-        with mock.patch("wandas.processing.create_operation") as mock_create_op:
-            # モックSTFTオペレーションの設定
-            mock_stft = mock.MagicMock(spec=STFT)
-            mock_data = mock.MagicMock(spec=DaArray)
-            mock_data.ndim = 3  # Set ndim property to pass dimension check
-            mock_data.shape = (2, 1025, 10)  # Set appropriate shape for a 3D array
-            mock_stft.process.return_value = mock_data
-            mock_create_op.return_value = mock_stft
-
-            # stftを遅延実行（デフォルト引数）
-            result = self.channel_frame.stft()
-
-            # デフォルトパラメータの確認
-            mock_create_op.assert_called_with(
-                "stft",
-                self.sample_rate,
-                n_fft=2048,
-                hop_length=512,  # n_fft//4
-                win_length=2048,
-                window="hann",
-            )
-
-            # processメソッドが呼び出されたか確認
-            mock_stft.process.assert_called_once_with(self.channel_frame._data)
-
-            # 結果が正しい型か確認
-            assert isinstance(result, SpectrogramFrame)
-            assert result.n_fft == 2048
-            assert result.hop_length == 512
-            assert result.win_length == 2048
-            assert result.window == "hann"
-
-            # カスタムパラメータでテスト
-            mock_create_op.reset_mock()
-            mock_stft.process.reset_mock()
-
-            # Update mock data shape for n_fft=1024
-            mock_data.shape = (
-                2,
-                513,
-                10,
-            )  # For n_fft=1024, freq_bins = 1024 // 2 + 1 = 513
-
-            result = self.channel_frame.stft(
-                n_fft=1024,
-                hop_length=256,
-                win_length=1024,
-                window="hamming",
-            )
-
-            mock_create_op.assert_called_with(
-                "stft",
-                self.sample_rate,
-                n_fft=1024,
-                hop_length=256,
-                win_length=1024,
-                window="hamming",
-            )
-
-            assert result.n_fft == 1024
-            assert result.hop_length == 256
-            assert result.win_length == 1024
-            assert result.window == "hamming"
-
-    def test_noct_spectrum_transform(self) -> None:
-        """Test noct_spectrum method for calculating N-octave spectrum analysis."""
-        from wandas.frames.noct import NOctFrame
-        from wandas.processing import NOctSpectrum
-
-        with mock.patch("wandas.processing.create_operation") as mock_create_op:
-            # モックNOctSpectrumオペレーションの設定
-            mock_noct = mock.MagicMock(spec=NOctSpectrum)
-            mock_data = mock.MagicMock(spec=DaArray)
-            mock_data.ndim = 2
-            mock_data.shape = (2, 10)  # バンド数に応じた適切な形状
-            mock_noct.process.return_value = mock_data
-            mock_create_op.return_value = mock_noct
-
-            # noct_spectrumを呼び出す
-            fmin, fmax, n = 20, 20000, 3
-            G, fr = 10, 1000  # noqa: N806
-            result = self.channel_frame.noct_spectrum(
-                fmin=fmin, fmax=fmax, n=n, G=G, fr=fr
-            )
-
-            # オペレーションが正しく作成されたか確認
-            mock_create_op.assert_called_with(
-                "noct_spectrum",
-                self.sample_rate,
-                fmin=fmin,
-                fmax=fmax,
-                n=n,
-                G=G,
-                fr=fr,
-            )
-
-            # processメソッドが呼び出されたか確認
-            mock_noct.process.assert_called_once_with(self.channel_frame._data)
-
-            # 結果が正しい型か確認
-            assert isinstance(result, NOctFrame)
-            assert result.fmin == fmin
-            assert result.fmax == fmax
-            assert result.n == n
-            assert result.G == G
-            assert result.fr == fr
-            assert result.previous is self.channel_frame
 
     def test_describe_method(self) -> None:
         """Test the describe method for visual and audio display."""
@@ -1033,9 +643,6 @@ class TestChannelFrame:
             assert mock_audio.call_count == self.channel_frame.n_channels
             for call in mock_audio.call_args_list:
                 assert call[1].get("normalize", True) is True
-
-            # モックのclose呼び出し回数はチャンネル数と完全に一致するとは限らない
-            # 実装が変更されている可能性があるため、呼び出し回数のチェックを削除
 
     def test_describe_method_with_axis_config(self) -> None:
         """Test describe method with legacy axis_config parameter."""
@@ -1100,370 +707,71 @@ class TestChannelFrame:
             with pytest.raises(TypeError, match="Unexpected type for plot result"):
                 self.channel_frame.describe()
 
-    def test_trim(self) -> None:
-        """Test the trim method."""
-        # Test trimming with start and end times
-        trimmed_frame = self.channel_frame.trim(start=0.1, end=0.5)
-        assert isinstance(trimmed_frame, ChannelFrame)
-        assert trimmed_frame.n_samples == int(0.4 * self.sample_rate)
-        assert trimmed_frame.n_channels == self.channel_frame.n_channels
+    def test_read_csv(self) -> None:
+        """Test read_csv method."""
+        # Create a temporary CSV file
+        with tempfile.NamedTemporaryFile(suffix=".csv", delete=False) as temp_file:
+            temp_filename = temp_file.name
+            # 時間と2つの値を持つCSVファイルを作成
+            header = "time,value1,value2\n"
+            data = "\n".join([f"{i / 16000},{1.1},{2.2}" for i in range(16000)])
+            temp_file.write(header.encode())
+            temp_file.write(data.encode())
+            temp_file.flush()
+            temp_file.seek(0)
+            # Close the file to ensure it's written
+            temp_file.close()
 
-        # Test trimming with only start time
-        trimmed_frame = self.channel_frame.trim(start=0.2)
-        assert isinstance(trimmed_frame, ChannelFrame)
-        assert trimmed_frame.n_samples == int(0.8 * self.sample_rate)
+        try:
+            # Read the CSV file into a ChannelFrame
+            cf = ChannelFrame.read_csv(temp_filename)
 
-        # Test trimming with only end time
-        trimmed_frame = self.channel_frame.trim(end=0.3)
-        assert isinstance(trimmed_frame, ChannelFrame)
-        assert trimmed_frame.n_samples == int(0.3 * self.sample_rate)
+            # Check properties
+            assert cf.sampling_rate == 16000
+            assert cf.n_channels == 2
+            assert cf.n_samples == 16000
+            assert cf.label == Path(temp_filename).stem
 
-        # Test trimming with no start or end (should return the same frame)
-        trimmed_frame = self.channel_frame.trim()
-        assert isinstance(trimmed_frame, ChannelFrame)
-        assert trimmed_frame.n_samples == self.channel_frame.n_samples
+            # Check data
+            expected_data = np.loadtxt(temp_filename, delimiter=",", skiprows=1).T
+            np.testing.assert_array_equal(cf.data, expected_data[1:])
 
-        # Test trimming with invalid start and end times
-        with pytest.raises(ValueError):
-            self.channel_frame.trim(start=0.5, end=0.1)
+        finally:
+            if os.path.exists(temp_filename):
+                os.remove(temp_filename)
 
-    def test_add_method(self) -> None:
-        """Test add method for adding signals."""
-        # 通常の加算をテスト
-        # Create another ChannelFrame
-        other_data = np.random.random((2, 16000))
-        other_dask_data = _da_from_array(other_data, chunks=-1)
-        other_cf = ChannelFrame(other_dask_data, self.sample_rate, label="other_audio")
+    def test_debug_info(self) -> None:
+        """Test debug_info method."""
+        with mock.patch("wandas.core.base_frame.logger") as mock_logger:
+            self.channel_frame.debug_info()
+            assert mock_logger.debug.call_count >= 6  # At least 6 debug messages
 
-        # addメソッドを使用して加算
-        result = self.channel_frame.add(other_cf)
+    def test_read_wav_class_method(self) -> None:
+        """Test read_wav class method."""
+        with mock.patch.object(ChannelFrame, "from_file") as mock_from_file:
+            mock_from_file.return_value = self.channel_frame
+            result = ChannelFrame.read_wav("test.wav", labels=["left", "right"])
+            mock_from_file.assert_called_with("test.wav", ch_labels=["left", "right"])
+            assert result is self.channel_frame
 
-        # Check result properties
-        assert isinstance(result, ChannelFrame)
-        assert result.sampling_rate == self.sample_rate
-        assert result.n_channels == 2
-        assert result.n_samples == 16000
+    def test_visualize_graph(self) -> None:
+        """Test visualize_graph method."""
+        # Test successful visualization
+        with mock.patch.object(DaArray, "visualize") as mock_visualize:
+            filename = self.channel_frame.visualize_graph()
+            mock_visualize.assert_called_once()
+            assert filename is not None
 
-        # 実際の計算結果を確認
-        computed = result.compute()
-        expected = self.data + other_data
-        np.testing.assert_array_almost_equal(computed, expected)
+        # Test with provided filename
+        with mock.patch.object(DaArray, "visualize") as mock_visualize:
+            custom_filename = "test_graph.png"
+            filename = self.channel_frame.visualize_graph(filename=custom_filename)
+            mock_visualize.assert_called_with(filename=custom_filename)
+            assert filename == custom_filename
 
-        # スカラー値との加算をテスト
-        scalar_value = 0.5
-        result = self.channel_frame.add(scalar_value)
-        assert isinstance(result, ChannelFrame)
-        computed = result.compute()
-        expected = self.data + scalar_value
-        np.testing.assert_array_almost_equal(computed, expected)
-
-        # NumPy配列との加算をテスト
-        array_value = np.random.random((2, 16000))
-        result = self.channel_frame.add(array_value)
-        assert isinstance(result, ChannelFrame)
-        computed = result.compute()
-        expected = self.data + array_value
-        np.testing.assert_array_almost_equal(computed, expected)
-
-        # サンプリングレートが一致しない場合のエラーをテスト
-        mismatch_cf = ChannelFrame(other_dask_data, 44100, label="mismatch_audio")
-        with pytest.raises(
-            ValueError, match="Sampling rates do not match. Cannot perform operation."
+        # Test handling of visualization error
+        with mock.patch.object(
+            DaArray, "visualize", side_effect=Exception("Test error")
         ):
-            _ = self.channel_frame.add(mismatch_cf)
-
-    def test_add_with_snr(self) -> None:
-        """Test add method with SNR parameter."""
-        # 別のChannelFrameを作成
-        signal_data = np.random.random((2, 16000))
-        signal_dask_data = _da_from_array(signal_data, chunks=-1)
-        signal_cf = ChannelFrame(signal_dask_data, self.sample_rate, label="signal")
-
-        # ノイズデータを作成
-        noise_data = np.random.random((2, 16000)) * 0.1  # 小さいノイズ
-        noise_dask_data = _da_from_array(noise_data, chunks=-1)
-        noise_cf = ChannelFrame(noise_dask_data, self.sample_rate, label="noise")
-
-        # SNRを指定して加算
-        snr_value = 10.0  # 10dBのSNR
-        result = signal_cf.add(noise_cf, snr=snr_value)
-
-        # 基本的なプロパティをチェック
-        assert isinstance(result, ChannelFrame)
-        assert result.sampling_rate == self.sample_rate
-        assert result.n_channels == 2
-        assert result.n_samples == 16000
-
-        # 演算履歴の確認 - 実装に合わせて調整
-        # この部分はapply_addの実装によって異なる可能性があるため、
-        # 一般的な作成チェックのみを行う
-        assert len(result.operation_history) > len(signal_cf.operation_history)
-
-        # 実際の計算をトリガー
-        computed = result.compute()
-
-        # SNRを考慮した加算の結果を確認
-        # 実際の結果はSNRの具体的な実装によって異なりますが、型と形状は確認可能
-        assert isinstance(computed, np.ndarray)
-        assert computed.shape == (2, 16000)
-
-        # 負のSNR値もテスト
-        # 値が適用されることを確認する
-        neg_result = signal_cf.add(noise_cf, snr=-10.0)
-        neg_computed = neg_result.compute()
-        assert isinstance(neg_computed, np.ndarray)
-        assert neg_computed.shape == (2, 16000)
-
-    def test_add_with_different_lengths(self) -> None:
-        """異なる長さの信号を加算するテスト。"""
-        # 標準の長さのフレーム（self.channel_frame）
-        # 長さが標準フレームよりも短いフレーム（切り詰め必要）
-        short_data = np.random.random((2, 8000))  # 半分の長さ
-        short_dask_data = _da_from_array(short_data, chunks=(1, 2000))
-        short_cf = ChannelFrame(short_dask_data, self.sample_rate, label="short_audio")
-
-        # 長さが標準フレームよりも長いフレーム（パディング必要）
-        long_data = np.random.random((2, 24000))  # 1.5倍の長さ
-        long_dask_data = _da_from_array(long_data, chunks=(1, 6000))
-        long_cf = ChannelFrame(long_dask_data, self.sample_rate, label="long_audio")
-
-        # 短いフレームを標準フレームに加算（パディングが必要）
-        result_short = self.channel_frame.add(short_cf)
-        computed_short = result_short.compute()
-
-        # 結果の形状が元のフレームと同じであることを確認
-        assert computed_short.shape == self.data.shape
-
-        # 短いフレーム部分は加算され、残りは元のフレームのままであることを確認
-        expected_short = self.data.copy()
-        expected_short[:, : short_data.shape[1]] = (
-            expected_short[:, : short_data.shape[1]] + short_data
-        )
-        np.testing.assert_array_almost_equal(computed_short, expected_short)
-
-        # 長いフレームを標準フレームに加算（切り詰めが必要）
-        result_long = self.channel_frame.add(long_cf)
-        computed_long = result_long.compute()
-
-        # 結果の形状が元のフレームと同じであることを確認
-        assert computed_long.shape == self.data.shape
-
-        # 元のフレームと同じ長さだけ長いフレームを切り詰めて加算されることを確認
-        expected_long = self.data + long_data[:, : self.data.shape[1]]
-        np.testing.assert_array_almost_equal(computed_long, expected_long)
-
-        # スカラー値や同じ長さの配列との加算は既に他のテストでカバーされているので省略
-
-    def test_csd(self) -> None:
-        """クロススペクトル密度（CSD）メソッドのテスト"""
-        # テスト用信号を作成
-        sr = 1000
-        t = np.linspace(0, 1, sr, endpoint=False)
-
-        # 既知の周波数成分を持つ信号を作成（100Hzと200Hz）
-        sig1 = np.sin(2 * np.pi * 100 * t) + 0.5 * np.sin(2 * np.pi * 200 * t)
-        sig2 = 0.8 * np.sin(2 * np.pi * 100 * t) + 0.3 * np.sin(2 * np.pi * 300 * t)
-
-        # 2チャンネル信号を作成
-        sigs = np.array([sig1, sig2])
-
-        # ChannelFrameインスタンスを作成
-        cf = ChannelFrame.from_numpy(sigs, sr)
-
-        # CSD計算のパラメータ
-        n_fft = 512
-        win_length = 256
-        hop_length = 128
-
-        # ChannelFrameメソッドを使用してCSDを計算
-        csd_frame = cf.csd(
-            n_fft=n_fft, win_length=win_length, hop_length=hop_length, window="hamming"
-        )
-
-        # 実際のデータを取得するために計算
-        csd_data = csd_frame.compute()
-
-        # 形状を確認
-        assert csd_data.shape == (4, n_fft // 2 + 1)
-
-        # チャンネルラベルが正しいことを確認
-        ch_pairs = [ch.label for ch in csd_frame._channel_metadata]
-
-        # 正しいラベル名を取得
-        ch0_label = cf.channels[0].label
-        ch1_label = cf.channels[1].label
-
-        expected_pairs = [
-            f"csd({ch0_label}, {ch0_label})",
-            f"csd({ch0_label}, {ch1_label})",
-            f"csd({ch1_label}, {ch0_label})",
-            f"csd({ch1_label}, {ch1_label})",
-        ]
-
-        for pair in expected_pairs:
-            assert pair in ch_pairs
-
-        # 主要な周波数成分の存在を確認
-        freq_bins = np.fft.rfftfreq(n_fft, d=1 / sr)
-        idx_100hz = np.argmin(np.abs(freq_bins - 100))
-        idx_200hz = np.argmin(np.abs(freq_bins - 200))
-        idx_300hz = np.argmin(np.abs(freq_bins - 300))
-
-        # 自己クロススペクトル密度では対応する周波数でピークが見られるはず
-        # ch0 (sig1) の自己CSDは100Hzと200Hzでピークを持つ
-        ch0_auto = csd_data[0]  # ch0 -> ch0
-        assert abs(ch0_auto[idx_100hz]) > abs(ch0_auto[idx_300hz])
-        assert abs(ch0_auto[idx_200hz]) > abs(ch0_auto[idx_300hz])
-
-        # ch1 (sig2) の自己CSDは100Hzと300Hzでピークを持つ
-        ch1_auto = csd_data[3]  # ch1 -> ch1
-        assert abs(ch1_auto[idx_100hz]) > abs(ch1_auto[idx_200hz])
-        assert abs(ch1_auto[idx_300hz]) > abs(ch1_auto[idx_200hz])
-
-    def test_transfer_function(self) -> None:
-        """伝達関数メソッドのテスト"""
-        # 単純なシステム用のテスト信号を作成
-        sr = 1000
-        t = np.linspace(0, 1, sr, endpoint=False)
-
-        # 入力信号
-        input_sig = np.sin(2 * np.pi * 100 * t) + 0.5 * np.sin(2 * np.pi * 200 * t)
-
-        # 出力信号（100Hzでゲイン2.0、200Hzでゲイン1.5を持つ）
-        # より現実的にするためにノイズを追加
-        output_sig = (
-            2 * np.sin(2 * np.pi * 100 * t)
-            + 0.75 * np.sin(2 * np.pi * 200 * t)
-            + 0.05 * np.random.randn(len(t))
-        )
-
-        # 2チャンネル信号を作成
-        sigs = np.array([input_sig, output_sig])
-
-        # ChannelFrameインスタンスを作成
-        cf = ChannelFrame.from_numpy(sigs, sr)
-
-        # 伝達関数計算のパラメータ
-        n_fft = 512
-        win_length = 256
-        hop_length = 128
-
-        # ChannelFrameメソッドを使用して伝達関数を計算
-        tf_frame = cf.transfer_function(
-            n_fft=n_fft, win_length=win_length, hop_length=hop_length, window="hamming"
-        )
-
-        # 実際のデータを取得するために計算
-        tf_data = tf_frame.compute()
-
-        # 形状を確認
-        assert tf_data.shape == (4, n_fft // 2 + 1)
-
-        # テスト周波数のインデックスを探す
-        freq_bins = np.fft.rfftfreq(n_fft, 1 / sr)
-        idx_100hz = np.argmin(np.abs(freq_bins - 100))
-        idx_200hz = np.argmin(np.abs(freq_bins - 200))
-
-        # 入力から出力への伝達関数（インデックス1は入力->出力）
-        # 主要周波数でのゲインがほぼ正確かチェック
-        h_in_to_out = tf_data[1]
-        assert np.isclose(np.abs(h_in_to_out[idx_100hz]), 2.0, rtol=0.2)
-        assert np.isclose(np.abs(h_in_to_out[idx_200hz]), 1.5, rtol=0.2)
-
-        # 自己伝達関数は約1.0になるはず
-        assert np.isclose(np.abs(tf_data[0, idx_100hz]), 1.0, rtol=0.2)  # 入力->入力
-        assert np.isclose(np.abs(tf_data[3, idx_100hz]), 1.0, rtol=0.2)  # 出力->出力
-
-        # チャンネルラベルが正しいことを確認
-        ch_pairs = [ch.label for ch in tf_frame._channel_metadata]
-
-        # 正しいラベル名を取得
-        ch0_label = cf.channels[0].label
-        ch1_label = cf.channels[1].label
-
-        expected_pairs = [
-            f"$H_{{{ch0_label}, {ch0_label}}}$",
-            f"$H_{{{ch0_label}, {ch1_label}}}$",
-            f"$H_{{{ch1_label}, {ch0_label}}}$",
-            f"$H_{{{ch1_label}, {ch1_label}}}$",
-        ]
-
-        for pair in expected_pairs:
-            assert pair in ch_pairs
-
-    def test_coherence(self) -> None:
-        """コヒーレンスメソッドのテスト"""
-        # テスト用信号を作成
-        sr = 1000
-        t = np.linspace(0, 1, sr, endpoint=False)
-
-        # 関連性のある2つの信号を作成
-        sig1 = np.sin(2 * np.pi * 100 * t) + 0.5 * np.sin(2 * np.pi * 200 * t)
-        sig2 = (
-            0.7 * np.sin(2 * np.pi * 100 * t)
-            + 0.4 * np.sin(2 * np.pi * 200 * t)
-            + 0.3 * np.sin(2 * np.pi * 300 * t)
-        )
-
-        # 2チャンネル信号を作成
-        sigs = np.array([sig1, sig2])
-
-        # ChannelFrameインスタンスを作成
-        cf = ChannelFrame.from_numpy(sigs, sr)
-
-        # コヒーレンス計算のパラメータ
-        n_fft = 512
-        win_length = 256
-        hop_length = 128
-
-        # ChannelFrameメソッドを使用してコヒーレンスを計算
-        coherence_frame = cf.coherence(
-            n_fft=n_fft, win_length=win_length, hop_length=hop_length, window="hamming"
-        )
-
-        # 実際のデータを取得するために計算
-        coherence_data = coherence_frame.compute()
-
-        # 形状を確認
-        assert coherence_data.shape == (4, n_fft // 2 + 1)
-
-        # 主要な周波数bins
-        freq_bins = np.fft.rfftfreq(n_fft, d=1 / sr)
-        idx_100hz = np.argmin(np.abs(freq_bins - 100))
-        idx_300hz = np.argmin(np.abs(freq_bins - 300))
-
-        # コヒーレンスの範囲チェック（0～1）
-        # 数値計算の誤差を考慮して、わずかに1を超える値も許容する
-        assert np.all(coherence_data >= 0)
-        assert np.all(coherence_data <= 1.001)  # 許容誤差を追加
-
-        # 自己コヒーレンスは1.0に近いはず
-        assert np.isclose(
-            coherence_data[0, idx_100hz], 1.0, rtol=1e-5
-        )  # チャンネル0自己コヒーレンス
-        assert np.isclose(
-            coherence_data[3, idx_100hz], 1.0, rtol=1e-5
-        )  # チャンネル1自己コヒーレンス
-
-        # チャンネル間のコヒーレンスを100Hzと300Hzで確認
-        # 100Hzでは両方の信号に成分があるので高いコヒーレンス
-        # 300Hzではチャンネル2にのみ成分があるので低いコヒーレンス
-        assert coherence_data[1, idx_100hz] > 0.8  # 100Hzでの高いコヒーレンス
-        assert coherence_data[1, idx_300hz] < 0.5  # 300Hzでの低いコヒーレンス
-
-        # チャンネルラベルが正しいことを確認
-        ch_pairs = [ch.label for ch in coherence_frame._channel_metadata]
-
-        # 正しいラベル名を取得
-        ch0_label = cf.channels[0].label
-        ch1_label = cf.channels[1].label
-
-        expected_pairs = [
-            f"$\\gamma_{{{ch0_label}, {ch0_label}}}$",
-            f"$\\gamma_{{{ch0_label}, {ch1_label}}}$",
-            f"$\\gamma_{{{ch1_label}, {ch0_label}}}$",
-            f"$\\gamma_{{{ch1_label}, {ch1_label}}}$",
-        ]
-
-        for pair in expected_pairs:
-            assert pair in ch_pairs
+            filename = self.channel_frame.visualize_graph()
+            assert filename is None
